pyproject.toml
[project]
name = "kanue"
version = "0.1.0"
description = "KAN-based chess evaluation: replacing NNUE activations with trainable Kolmogorov-Arnold Networks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "y0sif" }]
keywords = ["chess", "nnue", "kan", "kolmogorov-arnold", "neural-network", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Games/Entertainment :: Board Games",
]

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "python-chess>=1.9",
    "matplotlib>=3.7",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "nbstripout>=0.7",
]
colab = [
    "efficient-kan",
]

[project.urls]
Repository = "https://github.com/y0sif/kanue"
Issues = "https://github.com/y0sif/kanue/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[tool.hatch.build.targets.wheel]
packages = ["src/kanue"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
